Fixed blood moon double hemogen drain with multiple maps, slight other changes
The current code was checking for the blood moon on the map the player is viewing, instead the affected pawn being on. It was possible to abuse it by switching maps (if the player had multiple active), or moving pawns out on a caravan and creating a new colony.

I've also changed a bit how the patch works - instead of being a postfix that directly modifies the amount of hemogen the pawn has, followed by calling GeneResourceDrainUtility.PostResourceOffset - it's now a prefix that just doubles the amount the original method works with. This prevents the double call to GeneResourceDrainUtility.PostResourceOffset, as the original method will already work on the correct amount.