Make hemogen drain gizmo tooltip show doubled drain during blood moons
I understand this is probably something rather rather insignificant that received a transpiler, so I wouldn't be surprised if it got rejected.

I've patched GeneGizmo_ResourceHemogen.GetTooltip to:
- Cache if the blood moon is active
- Replace the call to IGeneResourceDrain.ResourceLossPerDay with our custom one
- Inside of our custom replaced methods, we call IGeneResourceDrain.ResourceLossPerDay and return the value as doubled if the blood moon is active (based on the value that was cached earlier)

I've tried to follow the same style as in the rest of the project, however if something doesn't match it - let me know and I'll change it. Same with any additional changes, if any are required.